Rename bottom bar to footer & move it to homepage
